+++ src/main/java/sunsetsatellite/retrostorage/block/RecipeEncoderBlock.java
@@ -4,7 +4,9 @@ import net.danygames2014.nyalib.block.DropInventoryOnBreak;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.block.material.Material;
 import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.world.BlockView;
 import net.minecraft.world.World;
+import net.modificationstation.stationapi.api.world.BlockStateView;
 import sunsetsatellite.catalyst.Catalyst;
 import sunsetsatellite.catalyst.core.util.model.FullyRotatableBlockWithEntity;
 import sunsetsatellite.retrostorage.RetroStorage;
@@ -36,4 +38,8 @@ public class RecipeEncoderBlock extends FullyRotatableBlockWithEntity implements
         return true;
     }
 
+    @Override
+    public boolean renderLayer(BlockView view, BlockStateView blockStateView, int x, int y, int z, int meta, int layer) {
+        return layer == 0;
+    }
 }